requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0.0
python-jose[cryptography]==3.3.0
structlog==24.4.0
# для тестов в Docker
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
